feat: add getPreviewRef and clear the preview cookie on exit
Add `getPreviewRef`, the read-side counterpart to `redirectToPreviewURL`,
for passing the preview ref into a cached function under Cache Components.
`enableAutoPreviews` now delegates to it so the read logic stays in sync.

`exitPreview` now clears the Prismic preview cookie, closing the
preview-cookie loop (write -> read -> clear).
